BayesianShockEstimation

Documentation for BayesianShockEstimation.

BayesianShockEstimation.BoundaryParams

BayesianShockEstimation.ShockParams

BayesianShockEstimation.BackwardsSample

BayesianShockEstimation.BallVolume

BayesianShockEstimation.ComputeIntersection

BayesianShockEstimation.IncrementShockCounts!

BayesianShockEstimation.KalmanFilterWalk

BayesianShockEstimation.RTS_Smooth

BayesianShockEstimation.SampleBoundary

BayesianShockEstimation.ShockDensity

BayesianShockEstimation.BoundaryParams — Type

BoundaryParams

Struct containing all needed information for boundary sampling

Parameters

μ : Mean at sample times

σ : Standard deviation at sample times

t : Sample times

BayesianShockEstimation.ShockParams — Type

ShockParams

Struct containing all need information for deterministic shock simulations

Parameters

dynamics: Representation of the dynamics

positions: List of sample positions

ballsize: Integration ball size

threshold: Minimum shock threshold

dimension: System dimensionality

BayesianShockEstimation.BackwardsSample — Method

Helper Function computing backwards smoothing according to RTS

BayesianShockEstimation.BallVolume — Method

BallVolume(systemparams::ShockParams)::Float64

Helper function that computes the volume of the integration ball

BayesianShockEstimation.ComputeIntersection — Method

ComputeIntersection(t1, u1, t2, u2)

Helper function to compute the intersection between two characteristic curves in Burgers equation starting at x=0.

BayesianShockEstimation.IncrementShockCounts! — Method

IncrementShockCounts!(shockcounts, systemparams, boundary)

Computes shocks for a given boundary.

Arguments

shockcounts: Array of current shock counts

systemparams : Struct of parameters for system simulation

boundary: Boundary conditions being tested

Returns

A list of approximate densities

Notes

Uses the method of characteristics to find shocks.

BayesianShockEstimation.KalmanFilterWalk — Method

Helper function computing Kalman filter for the given model

BayesianShockEstimation.RTS_Smooth — Method

Helper Function computing variance for RTS sampler

Returns forward estimates with RTS variances

BayesianShockEstimation.SampleBoundary — Method

SampleBoundary(boundaryparams)

Generate a sample of a boundary.

Arguments

boundaryparams: Parameters characterizing the distribution of the boundary

Returns

A function representing a sample of boundary conditions.

BayesianShockEstimation.ShockDensity — Method

ShockDensity(systemparams, samples)

Generates an approximation of the shock density function.

Arguments

systemparams : Struct of parameters for system simulation

samples: Number of monte carlo samples

Returns

A list of approximate densities