BayesianShockEstimation

Documentation for BayesianShockEstimation.

BayesianShockEstimation.BoundaryParams

BayesianShockEstimation.ShockParams

BayesianShockEstimation.BackwardsVarianceWalk

BayesianShockEstimation.BallVolume

BayesianShockEstimation.ComputeIntersection

BayesianShockEstimation.FindPointsFromCurves

BayesianShockEstimation.IncrementShockCounts!

BayesianShockEstimation.KalmanFilterWalk

BayesianShockEstimation.RTS_Smooth

BayesianShockEstimation.RTS_sample

BayesianShockEstimation.ShockDensity

BayesianShockEstimation.distanceToSegment

BayesianShockEstimation.flux_burgers

BayesianShockEstimation.flux_difference!

BayesianShockEstimation.godunov_burgers_1D

BayesianShockEstimation.godunov_burgers_1D_step!

BayesianShockEstimation.riemann_solver_burgers

BayesianShockEstimation.riemann_solver_burgers!

BayesianShockEstimation.BoundaryParams — Type

BoundaryParams

Struct containing all needed information for boundary sampling

Parameters

σ_w : System noise standard deviation

σ_v : Measurement noise standard deviation

t : Sample interval

BayesianShockEstimation.ShockParams — Type

ShockParams

Struct containing all need information for deterministic shock simulations

Parameters

dynamics: Representation of the dynamics

stepsize: Grid size

ballsize: Integration ball size

threshold: Minimum shock threshold

dimension: System dimensionality

BayesianShockEstimation.BackwardsVarianceWalk — Method

Helper Function computing variance from RTS smoother

BayesianShockEstimation.BallVolume — Method

BallVolume(systemparams::ShockParams)::Float64

Helper function that computes the volume of the integration ball

BayesianShockEstimation.ComputeIntersection — Method

ComputeIntersection(t1, u1, t2, u2)

Helper function to compute the intersection between two characteristic curves in Burgers equation starting at x=0.

BayesianShockEstimation.FindPointsFromCurves — Method

Helper function that converts curves to indices to increment

BayesianShockEstimation.IncrementShockCounts! — Method

IncrementShockCounts!(shockcounts, systemparams, boundary)

Computes shocks for a given boundary.

Arguments

shockcounts: Array of current shock counts

systemparams : Struct of parameters for system simulation

T: Sampling Interval

boundary_values: Boundary samples

Returns

A list of approximate densities

Notes

Uses the method of characteristics to find shocks.

BayesianShockEstimation.KalmanFilterWalk — Method

Helper function computing Kalman filter for the given model

BayesianShockEstimation.RTS_Smooth — Method

Helper Function computing variance for RTS sampler

Returns forward estimates with RTS variances

BayesianShockEstimation.RTS_sample — Method

Sample from RTS smoother backwards sampler for random walk

BayesianShockEstimation.ShockDensity — Method

ShockDensity(systemparams, samples)

Generates an approximation of the shock density function.

Arguments

systemparams : Struct of parameters for system simulation

samples: Number of monte carlo samples

Returns

A list of approximate densities

BayesianShockEstimation.distanceToSegment — Method

Helper function computing distance from p3 to the line segment defined by the other points.

BayesianShockEstimation.flux_burgers — Method

flux_burgers(u)

Helper function returns flux in from the conservation form of Burgers' Equation.

Arguments

u - The state value

BayesianShockEstimation.flux_difference! — Method

flux_difference!(boundary_flux, boundary_flux_right)

Compute the difference between the right and left flux

Arguments

boundary_flux - output array

boundary_flux_right - flux through right boundaries

BayesianShockEstimation.godunov_burgers_1D — Method

godunov_burgers_1D(u0, dx, dt, stepcount)

Solve Burgers' equation numericallly with Godunov's method

Arguments

u0 - Initial Boundary Condition

dx - Spatial grid size

dt - Temporal grid size

stepcount - Number of steps

Returns

A 2D array representing the numerical solution of Burgers' Equation

BayesianShockEstimation.godunov_burgers_1D_step! — Method

godunov_burgers_1D_step!(u_next, boundary_flux, boundary_flux_right, u, dx, dt)

Step Burgers' equation according to Godunov's method

Arguments

u_next - Array to store the value

boundary_flux - Array of boundary flux storage

boundary_flux_right - Array of right boundary flux storage

u - Array of values for the previous timestep

dx - Spatial grid size

dt - Temporal grid size

Returns

An array of values for the next timestep

Notes

Assumes periodic boundary conditions

BayesianShockEstimation.riemann_solver_burgers! — Method

riemann_solver_burgers!(boundary_flux_right, u)

Solve for all right boundary flux values.

Arguments

boundary_flux_right - Output Array

u - State Values

BayesianShockEstimation.riemann_solver_burgers — Method

riemann_solver_burgers(ul, ur)

Solves the Riemann problem for Burgers' equation and returns the average value. Helper function for Godunov's scheme.

Arguments

ul - The left state value

`ur - The right state value

Returns

The value of flux on the boundary between the regions.