BayesianShockEstimation

Documentation for BayesianShockEstimation.

There are three main parts of the code for this project.

Pushing Densities into Shockwaves

Godunov scheme for numerically solving Burgers' equation

RTS Sampling of a random walk

The novelty in this work comes from the numerical computation of shockwave densities, while the other sections can be considered standard.

Shockwave Density

First, we introduce two structs for configuration of the simulation. The first focuses on the numerical details, while the second focuses on the boundary conditions.

BayesianShockEstimation.ShockParams — Type

ShockParams

Struct containing all needed information for deterministic shock simulations

Parameters

stepsize_x: Spatial grid size

stepsize_t: Temporal grid size

stepcount : Number of temporal steps

ballsize : Integration ball size

threshold : Minimum shock threshold

BayesianShockEstimation.BoundaryParams — Type

BoundaryParams

Struct containing all needed information for boundary sampling

Parameters

σ_w : System noise standard deviation

σ_v : Measurement noise standard deviation

Using these structs for configuration, the main function in this work computes the approximate shockwave arrival rate with a given number of Monte Carlo samples.

BayesianShockEstimation.ShockDensity — Function

ShockDensity(  systemparams::ShockParams, 
               boundaryparams::BoundaryParams, 
               observations, 
               samplecount)

Generates an approximation of the shock density function.

Arguments

systemparams : Struct of parameters for system simulation

boundaryparams : Struct of parameters for boundary simulation

observations : Boundary observations

samplecount : Number of Monte Carlo samples

Returns

A 2D array of approximate shockwave densities

Within the function we do in-place updates to count of shockwaves from Monte Carlo samples using IncrementShockCounts!.

BayesianShockEstimation.IncrementShockCounts! — Function

IncrementShockCounts!(shockcounts, systemparams, boundary)

Computes shocks for a given boundary.

Arguments

shockcounts : Array of current shock counts

systemparams : Struct of parameters for system simulation

boundary_values : Boundary samples

Returns

A list of approximate densities

The $\epsilon$ ball normalization term is computed using BallVolume.

Missing docstring.

Missing docstring for BallVolume. Check Documenter's build log for details.

Godunov Scheme for Burgers' Equation

To numerically compute the locations of shockwaves, we use Godunov's scheme for a discretization due to the balance of computational efficiency and the lack of advection error.

It is important to note that Gudonov's scheme introduces artificial viscosity into Burgers' equation, and so we only search for regions with large derivatives.

The highest level function for this section is godunov_burgers_1D_shocks, which returns a 2D array of booleans indicating the locations of shocks for given boundary conditions.

BayesianShockEstimation.godunov_burgers_1D_shocks — Function

godunov_burgers_1D_shocks(u0, dx, dt, stepcount, threshold)

Finds spatial steps in the numerical solution of Burgers equation with a slope greater than the threshold. These points represent our approximations of shocks.

Arguments

u0 : Initial Boundary Condition

dx : Spatial grid size

dt : Temporal grid size

stepcount : Number of steps

threshold : slope threshold

One step lower level is the numerical solver itself.

BayesianShockEstimation.godunov_burgers_1D — Function

godunov_burgers_1D(u0, dx, dt, stepcount)

Solve Burgers' equation numericallly with Godunov's method

Arguments

u0 : Initial Boundary Condition

dx : Spatial grid size

dt : Temporal grid size

stepcount : Number of steps

Returns

A 2D array representing the numerical solution of Burgers' Equation

Individual timesteps are completed using godunov_burgers_1D_step!, which operates on a pre-allocated array.

BayesianShockEstimation.godunov_burgers_1D_step! — Function

godunov_burgers_1D_step!(u_next, boundary_flux, boundary_flux_right, u, dx, dt)

Step Burgers' equation according to Godunov's method

Arguments

u_next : Array to store the value

boundary_flux : Array of boundary flux storage

boundary_flux_right : Array of right boundary flux storage

u : Array of values for the previous timestep

dx : Spatial grid size

dt : Temporal grid size

Returns

An array of values for the next timestep

Notes

Assumes periodic boundary conditions

The flux function for Burgers equation is represented as flux_burgers.

BayesianShockEstimation.flux_burgers — Function

flux_burgers(u)

Helper function returns flux in from the conservation form of Burgers' Equation.

Arguments

u : The state value

For any given boundary, the net flux is computed through an exact solution of the Riemann problem in riemann_solver_burgers.

BayesianShockEstimation.riemann_solver_burgers — Function

riemann_solver_burgers(ul, ur)

Solves the Riemann problem for Burgers' equation and returns the average value. Helper function for Godunov's scheme.

Arguments

ul : The left state value

ur : The right state value

Returns

The value of flux on the boundary between the regions.

Finally, there are two helper functions to deal with fused operations and periodic boundary conditions.

BayesianShockEstimation.riemann_solver_burgers! — Function

riemann_solver_burgers!(boundary_flux_right, u)

Solve for all right boundary flux values.

Arguments

boundary_flux_right : Output Array

u : State Values

BayesianShockEstimation.flux_difference! — Function

flux_difference!(boundary_flux, boundary_flux_right)

Compute the difference between the right and left flux

Arguments

boundary_flux : output array

boundary_flux_right : flux through right boundaries

RTS Sampler for the Random Walk Model

Because the shockwave arrival rate estimation arises by pushing the boundary density through an implicit transformation, we include an implementation of an RTS sampler to sample the conditional boundary distribution given observations under a random walk model.

The sampler makes use of RTS smoothing variances, Kalman filter estimates, and Kalman filter variances, all of which are generated by RTS_Smooth.

BayesianShockEstimation.RTS_Smooth — Function

RTS_Smooth(observations, σ_w, σ_v)

Helper Function computing variance for RTS sampler by computing the Kalman filter estimates as well as the RTS smoother variances.

Applies to the random walk model

\[\begin{aligned} x_{t+1} &= x_t + w_t\\ y_t &= x_t + v_t\\ \end{aligned}\]

where $w_t$ and $v_t$ are centered Gaussian random variables with variances $σ_w^2$ and $σ_v^2$, respectively.

Arguments

observations : Timeseries observation

σ_w : System noise standard deviation

σ_v : Observaton noise standard deviation

Returns

estimates, variances, RTSvariances

estimates : Kalman filter estimates

variances : Kalman filter variances

RTSvariances : RTS smoother variances

The results of RTS_Smooth can then be used in RTS_sample to generate a single trajectory sample according to the RTS smoothing distribution.

BayesianShockEstimation.RTS_sample — Function

RTS_sample(estimates, variances, RTSvariances, σ_w)

Sample from RTS smoother backwards sampler for random walk model.

Uses the Kalman filter estimates and variances, as well as the RTS variances generated by RTS_Smooth.

Arguments

estimates : Kalman filter estimates

variances : Kalman filter variances

RTSvariances : RTS smoother variances

σ_w : System noise standard deviation

The initial Kalman filter estimates are constructed through KalmanFilterWalk.

BayesianShockEstimation.KalmanFilterWalk — Function

KalmanFilterWalk(observations, σ_w, σ_v)

Computes the Kalman filter for a random walk model.

\[\begin{aligned} x_{t+1} &= x_t + w_t\\ y_t &= x_t + v_t\\ \end{aligned}\]

where $w_t$ and $v_t$ are centered Gaussian random variables with variances $σ_w^2$ and $σ_v^2$, respectively.

Arguments

observations : Timeseries observation

σ_w : System noise standard deviation

σ_v : Observaton noise standard deviation

Returns

estimates, variances

estimates : Kalman filter estimates

variances : Kalman filter variances

The RTS variances are constructed through BackwardsVarianceWalk.

BayesianShockEstimation.BackwardsVarianceWalk — Function

BackwardsVarianceWalk(variances, σ_w)

Given Kalman filter variances, computes the RTS smoother variances.

Arguments

variances : Kalman filter variances

σ_w : System noise power

Returns

The variances for the RTS smoother